Updated Unit test placement for select component
